+++ resources/css/app.css
@@ -18,14 +18,14 @@
     font-family: "DMSans-Italic";
     src: url("https://portal.davidici.com/fonts/DM_Sans/DMSans-Italic-VariableFont_opsz,wght.ttf");
 }
-/* @font-face {
-    font-family: "DMSans";
-    src: url("https://portal.davidici.com/fonts/DM_Sans/DMSans-VariableFont_opsz,wght.ttf");
-} */
 @font-face {
     font-family: "DMSans";
-    src: url("https://seven.test/fonts/DM_Sans/DMSans-VariableFont_opsz,wght.ttf");
+    src: url("https://portal.davidici.com/fonts/DM_Sans/DMSans-VariableFont_opsz,wght.ttf");
 }
+/* @font-face {
+    font-family: "DMSans";
+    src: url("https://seven.test/fonts/DM_Sans/DMSans-VariableFont_opsz,wght.ttf");
+} */
 /* @font-face {
     font-family: "DMSans";
     src: url("https://davidici-seven-portal.test/fonts/DM_Sans/DMSans-VariableFont_opsz,wght.ttf");
